- EMAR-60: API to get user information Emar.Api\Controllers\UsersController.cs Emar.Core\Users\Model\Mappings\UserMapper.cs Emar.Core\Users\Model\UserDto.cs Emar.Core\Users\Model\UserHeaderDto.cs Emar.Core\Users\Repository\UserRepository.cs Emar.Core\Users\Service\IUserService.cs Emar.Core\Users\Service\UserService.cs Emar.Data\EmarContext.cs Emar.Data\Entities\User.cs
- No ticket: Accommodate db changes
Emar.Core\Orders\Model\Mappings\OrderMapper.cs
Emar.Core\Orders\Model\OrderBase.cs
Emar.Core\Sites\Model\Mappings\SiteMapper.cs
Emar.Core\Sites\Model\SiteDto.cs
Emar.Data\Entities\OrderAdministration.cs
Emar.Data\Entities\OrderEvent.cs
Emar.Data\Entities\Patient.cs
Emar.Data\Entities\PatientOrder.cs
Emar.Data\Entities\Site.cs

- No ticket: Add the site name in the patient record
Emar.Core\Patients\Model\Mappings\PatientMapper.cs
Emar.Core\Patients\Model\PatientDto.cs
Emar.Core\Patients\Repository\PatientRepository.cs
Emar.Data\EmarContext.cs
Emar.Data\Entities\Patient.cs

- No ticket: Return patient orders in patient data
Emar.Core\Patients\Repository\PatientRepository.cs

- No ticket: Not all orders returned
Emar.Core\Orders\Repository\OrderRepository.cs

git-tfs-id: [http://ros-tfs-px01.picis.com:8080/tfs/emar]$/;C188

+++ WebAPI/API/C_Development/C_DevTrunk/Emar.Data/Entities/OrderEvent.cs
@@ -7,16 +7,13 @@ namespace Emar.Data.Entities
     [Table("order_events")]
     public class OrderEvent
     {
-        //[Column("id", TypeName = "bigint"), Key]
-        [Column("id", TypeName = "int"), Key]
+        [Column("id", TypeName = "bigint"), Key]
         public long Id { get; set; }
 
-        //[Column("patient_order_id", TypeName = "bigint"), Required]
-        [Column("patient_order_id", TypeName = "int"), Required]
+        [Column("patient_order_id", TypeName = "bigint"), Required]
         public long OrderId { get; set; }
 
-        //[Column("order_administration_id", TypeName = "bigint")]
-        [Column("order_administration_id", TypeName = "int")]
+        [Column("order_administration_id", TypeName = "bigint")]
         public long? AdministrationId { get; set; }
 
         [Column("event_datetime", TypeName = "datetimeoffset"), Required]
